migration: Add hash cache for dirty page tracking
Add a scheme that's complemetary to the Dirty bit (via the KVM log)
to represent has really change or not. The dirty bit provides with
hugepage areas that have dirtied pages, and with this scheme we extend
zero page detection in order to detect zeroes, yes, and in addition
whether a certain base page has changed or not. Such change occurs when
an entry in the hash cache has a different hash then what's stored in
the cache thus making it eligible to be sent on the wire.

The hash occurs on a 'live' page accessing Qemu map of guest RAM,
whereby when we have an event where the hash calculation between live and cache
is mismatched then a copy of the live page is performed, hashed, and
said page copy is sent to the destination. On a scheme that relies on
page contents, hash integrity must have assurance that it matches the
content of the RAM pages *in the destination*. This is purely from a
maintenance perspective, because we would otherwise have to proactively
invalidate the contents of the cache after we send *some state* of a
page to destination given that page can change between computation and
 being sent (and even after until we check again if page got change as
it went in the wire). So for simplicity we copy (and bear the
performance cost) and it simplifies the code around it. If there's an
entry of a marked zero page in the hash cache we must invalidate it
because the host will write zeroes on the destination.

The cache is used to store the copy of page to be sent over to the
destination. Because the cached page is scratch space it can't be hold
on forever; and not flushing it makes the next block of pages corrupt
said copy as it is still being sent on the wire. Instead let's make
zerocopy incompatible with hash cache migration just like we do with
device state. In terms of error handling, special emphasis on migration
failure if the hash cache allocation fails: it means we are out of memory
and we can't proceed ahead with the migration.

Convergence with this zerocopy method then is bound to how many pages we
need to hash before we check if we can converge. This is as opposed to
just 'outstanding' pages that we still have to be sent over, which we
A/D bit tracking will be at hugepage granularity and therefore a much
bigger super set.

Note: Migration will fail when using migrate-use-hash=on given the lack
of an algorithm. That is added after.

Orabug: 37699414
Co-developed-by: Joao Martins <joao.m.martins@oracle.com>
Signed-off-by: Joao Martins <joao.m.martins@oracle.com>
Signed-off-by: Elena Ufimtseva <elena.ufimtseva@oracle.com>
Reviewed-by: Boris Ostrovsky <boris.ostrovsky@oracle.com>