migration: Sync dirty bitmap in exact() considering factor
Sync the KVM bitmap not only on remaining being in threshold-size
boundaries, but also when the calculated factor based remaining is
within capacity.

Such factor is 1 without hash cache migration so this is a NOP for
regular write-protection based live migration. It really only matters
for hash cache migration.

Orabug: 37699414
Signed-off-by: Joao Martins <joao.m.martins@oracle.com>
Reviewed-by: Boris Ostrovsky <boris.ostrovsky@oracle.com>