migration/ram: Adjust estimate/exact to time-to-hash dirty data
Draw a downtime estimation against how much time we take to hash the
outstanding RAM instead of blindly taking the previous factor at face
value.

This prevents an overoptimistic scenario which does not consider that it
takes time to hash data which is not really captured in the algorithm.
If we are still within the boundaries of downtime, when we apply the
factor to the remaining size we return to the migration main thread
loop.

Orabug: 37699414
Signed-off-by: Joao Martins <joao.m.martins@oracle.com>
Reviewed-by: Boris Ostrovsky <boris.ostrovsky@oracle.com>