target/i386/kvm: account blackout downtime for kvm-clock and guest TSC
So far, QEMU/KVM live migration does not account all elapsed blackout
downtimes. For example, if a guest is live-migrated to a file, left idle
for one hour, and then restored from that file to the target host, the
one-hour blackout period will not be reflected in the kvm-clock or guest
TSC.

Typically, the elapsed time between KVM_GET_CLOCK (on the source QEMU) and
KVM_SET_CLOCK (on the target QEMU) is not accounted in the kvm-clock.
Similarly, the elapsed time between reading MSR_IA32_TSC on the source QEMU
and writing it on the target QEMU is not reflected in the guest TSC.

The KVM patchset [1] introduced KVM_VCPU_TSC_CTRL, KVM_CLOCK_REALTIME, and
KVM_CLOCK_HOST_TSC to account the elapsed time during live migration
blackouts in the guest's system counter view.

The core idea is to use the realtime clock (KVM_CLOCK_REALTIME) from both
the source and target hosts as a reference to calculate the elapsed
downtime in nanoseconds and adjust kvm-clock. In addition, these
nanoseconds are converted into TSC cycles using the vCPU's virtual TSC
frequency, and the corresponding elapsed cycles are compensated in the
guest TSC as well (KVM_CLOCK_HOST_TSC and KVM_VCPU_TSC_CTRL).

The following steps are copied from the Linux kernel documentation [2].

From the source VMM process:

1. Invoke the KVM_GET_CLOCK ioctl to record the host TSC (tsc_src),
kvmclock nanoseconds (guest_src), and host CLOCK_REALTIME nanoseconds
(host_src).

2. Read the KVM_VCPU_TSC_OFFSET attribute for every vCPU to record the
guest TSC offset (ofs_src[i]).

3. Invoke the KVM_GET_TSC_KHZ ioctl to record the frequency of the guest's
TSC (freq).

From the destination VMM process:

4. Invoke the KVM_SET_CLOCK ioctl, providing the source nanoseconds from
kvmclock (guest_src) and CLOCK_REALTIME (host_src) in their respective
fields. Ensure that the KVM_CLOCK_REALTIME flag is set in the provided
structure.

KVM will advance the VM's kvmclock to account for elapsed time since
recording the clock values. Note that this will cause problems in the guest
(e.g., timeouts) unless CLOCK_REALTIME is synchronized between the source
and destination, and a reasonably short time passes between the source
pausing the VMs and the destination executing steps 4-7.

5. Invoke the KVM_GET_CLOCK ioctl to record the host TSC (tsc_dest) and
kvmclock nanoseconds (guest_dest).

6. Adjust the guest TSC offsets for every vCPU to account for (1) time
elapsed since recording state and (2) difference in TSCs between the
source and destination machine:

ofs_dst[i] = ofs_src[i] -
    (guest_src - guest_dest) * freq + (tsc_src - tsc_dest)

("ofs[i] + tsc - guest * freq" is the guest TSC value corresponding to a
time of 0 in kvmclock. The above formula ensures that it is the same on
the destination as it was on the source).

7. Write the KVM_VCPU_TSC_OFFSET attribute for every vCPU with the
respective value derived in the previous step.

Introduce 'realtime', 'host_tsc', and 'flags' fields to KVMClockState,
and enable their live migration. Re-use CPUX86State->tsc_offset (currently
used by TCG) for the KVM accelerator.

References:
[1] https://lore.kernel.org/all/20210916181538.968978-8-oupton@google.com/
[2] https://docs.kernel.org/virt/kvm/devices/vcpu.html
[3] https://lore.kernel.org/qemu-devel/2d375ec3-a071-4ae3-b03a-05a823c48016@oracle.com/

Orabug: 38307402
Signed-off-by: Dongli Zhang <dongli.zhang@oracle.com>
Reviewed-by: Mark Kanda <mark.kanda@oracle.com>