migration: add parameter to specify max number of contiguous pages
that main thread will scan at a time.
Default value for non-exadata machine is 128.
It is the number that is currently used by live migration when
queueing pages to multifd threads.
For exadata machine type (unless hashing is disabled) the default
value for this parameter is 2048.
If hashing is disabled, the default value 128 is used for any
type of machine.

To change how many pages to scan during live migration at the time
for presence of contiguous block of pages, following can be used.

From HMP:
migrate_set_parameter migrate-scan-pages 4096

From Qemu command line:
-global migration.x-orcl-migrate-scan-pages=4096

From Libvirt XML:

<qemu:commandline>
        <qemu:arg value='-global'/>
        <qemu:arg value='migration.x-orcl-migrate-scan-pages=1024'/>
</qemu:commandline>

Orabug: 38388170

Signed-off-by: Elena Ufimtseva <elena.ufimtseva@oracle.com>
Reviewed-by: Maciej S. Szmigiero <maciej.szmigiero@oracle.com>