meson: check if isa-l installed and enable it
CONFIG_ISAL is enabled when the isa-l_crypto development package is
found at build time. Detection is done via pkg-config, which checks
for the presence of the .pc file. When CONFIG_ISAL is enabled, the
ISA-L crypto library name is set and the library SONAME major version
is passed to the configure script.

To preserve ABI compatibility of the isa-l_crypto library between
build time and runtime, the ABI major version is determined in the
RPM spec file. If the ABI version cannot be detected, a default value
of 2 is used in meson.build. The detected major version is then passed
to the QEMU configure script.

Orabug: 38732433

Signed-off-by: Elena Ufimtseva <elena.ufimtseva@oracle.com>
Reviewed-by: Maciej S. Szmigiero <maciej.szmigiero@oracle.com>