migration/page_cache: Add batching mode support for isa-l_crypto
This will be used in the next patch when adding to the backend. The
hash-based dirty tracking performs hashing of 4K pages one at a time,
and sorts 'non-zero pages' from 'dirty pages' based on matching hashes
in the QEMU page cache. The isa-l multi-buffer API, however, doesn't
perform well when we only use 1 lane (a.k.a. one hash computation at a
time). A huge speed-up is achieved when multi-buffers are submitted for
hashing (lanes in this library) in the same API. To leverage that speed,
introduce batching of hashes. It lets us perform the hashing all at once
(given that we know all of the pages are non-zero), and defer the matching
of the page cache stored hashes until the hashing is complete.

Orabug:  38732433

Signed-off-by: Joao Martins <joao.m.martins@oracle.com>
Co-developed-by: Elena Ufimtseva <elena.ufimtseva@oracle.com>
Signed-off-by: Elena Ufimtseva <elena.ufimtseva@oracle.com>
Reviewed-by: Boris Ostrovsky <boris.ostrovsky@oracle.com>